refactor matrix library on serveral files
