+++ src/BUILD
@@ -1,4 +1,45 @@
 cc_library(
+    name = "matrix",
+    srcs = [
+        "sp_matrix.c",
+        "dm_matrix.c",
+        "dm_math.c",
+        "dm_io.c",
+        "dv_vector.c",
+    ],
+    hdrs = [
+        "include/sp_matrix.h",
+        "include/dm_matrix.h",
+        "include/dm_math.h",
+        "include/dm_io.h",
+        "include/dv_vector.h",
+    ],
+    includes = ["include"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//lib/misc",
+        "@dbg-macro",
+    ],
+)
+
+
+cc_library(
+    name = "sp_matrix",
+    srcs = [
+        "sp_matrix.c",
+    ],
+    hdrs = [
+        "include/sp_matrix.h",
+    ],
+    includes = ["include"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//lib/misc",
+        "@dbg-macro",
+    ],
+)
+
+cc_library(
     name = "dm_matrix",
     srcs = [
         "dm_matrix.c",
@@ -10,6 +51,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//lib/misc",
+        "//src:sp_matrix",
         "@dbg-macro",
     ],
 )
@@ -26,6 +68,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//lib/misc",
+        "//src:sp_matrix",
         "//src:dm_matrix",
     ],
 )
@@ -42,6 +85,8 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         "//lib/misc",
+        "//src:sp_matrix",
         "//src:dm_matrix",
+        "//src:dm_math",
     ],
 )
