+++ src/include/dm_math.h
@@ -5,6 +5,8 @@
 #include "dm_matrix.h"
 #include "misc.h"
 
+
+
 /*******************************/
 /*      Double Matrix Math     */
 /*******************************/
